requirements.txt
requests
beautifulsoup4
scikit-learn
supabase
python-dotenv
streamlit